+++ src/frontend/web_application/src/components/Section/index.jsx
@@ -11,6 +11,7 @@ class Section extends PureComponent {
     descr: PropTypes.string,
     hasSeparator: PropTypes.bool,
     children: PropTypes.node,
+    borderContext: PropTypes.oneOf(['disabled', 'safe', 'public', 'unsecure']),
   };
 
   static defaultProps = {
@@ -19,15 +20,23 @@ class Section extends PureComponent {
     descr: undefined,
     hasSeparator: false,
     children: undefined,
+    borderContext: undefined,
   };
 
   render() {
     const {
-      title, descr, children, hasSeparator, className,
+      title, descr, children, hasSeparator, className, borderContext,
     } = this.props;
 
+    const borderModifier = borderContext && {
+      'm-section--border-disabled': borderContext === 'disabled',
+      'm-section--border-safe': borderContext === 'safe',
+      'm-section--border-public': borderContext === 'public',
+      'm-section--border-unsecure': borderContext === 'unsecure',
+    };
+
     return (
-      <section className={classnames('m-section', { 'm-section--separator': hasSeparator }, className)}>
+      <section className={classnames('m-section', { 'm-section--separator': hasSeparator }, borderModifier, className)}>
         {(title || descr) &&
           <header className="m-section__header">
             {title && <Title className="m-section__title">{title}</Title>}