+++ src/frontend/web_application/src/modules/remoteIdentity/actions/deleteRemoteIdentity.js
@@ -1,0 +1,8 @@
+import { deleteRemoteIdentity as deleteRemoteIdentityBase, removeFromCollection } from '../../../store/modules/remote-identity';
+import { tryCatchAxiosAction } from '../../../services/api-client';
+
+export const deleteRemoteIdentity = ({ remoteIdentity }) => async (dispatch) => {
+  await tryCatchAxiosAction(() => dispatch(deleteRemoteIdentityBase({ remoteIdentity })));
+
+  return dispatch(removeFromCollection({ remoteIdentity }));
+};