+++ src/frontend/web_application/src/modules/remoteIdentity/index.js
@@ -1,0 +1,2 @@
+export * from './actions/createRemoteIdentity';
+export * from './actions/deleteRemoteIdentity';